+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.6] - 2022-11-10
+### Fixed
+- Various bugs with TableauxSystem's `is_correct_tree` (no missing premises -avoids
+  marking tableaux that only contain one premise as correct-; all premises must come at 
+  the beginning of the tableaux, before branching and before applying any other rule)
+
 ## [1.1.5] - 2022-11-09
 ### Changed
 - TableauxSystem's `is_correct_tree` method now takes an extra `parser` optional parameter